// Code/Websites/DanpheEMR/wwwroot/DanpheApp/src/app/clinical-new/shared/dto/get-cln-vitals.dto.ts
export class GetClnVitals_DTO {
    VitalsId: number = 0;
    VitalsCode: string = "";
    VitalsType: string = "";
    VitalsName: string = "";
    VitalsValue: string = "";
    Unit: string = "";
    DisplayOrder: number = 0;
    VitalsGroup: string = "";
    CreatedBy: number = 0;
    CreatedOn: string = "";
    ModifiedOn: string = "";
    ModifiedBy: number = 0;
    IsActive: boolean = false;
    InputType: string = "";
}